Fixed issue #6
Fixed issue #6 where models who you must be logged in to see in the directory would not record. Login is not required (I dont have an account, and Im not willing to enter my credit card info to create one, so I can not reverse engineer the login process), the script will check each model in the wanted list at startup, and again every 4 hours to see if any of the models in the wanted list needs to checked manually rather than just through the online models json response.